requirements.txt
python-dotenv
openai
langchain==0.0.187
fire
gradio
tiktoken